pyproject.toml
[project]
name = "llama-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-llms-anthropic>=0.7.6",
    "llama-index-llms-groq>=0.3.2",
    "llama-index-llms-openai>=0.4.7",
    "openai>=1.97.1",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
